extract labelscreen list components to /diary/list
To tidy up /js/diary/, I am reorganizing it into subdirectories.
