+++ www/js/diary/list/TimelineScrollList.tsx
@@ -3,14 +3,14 @@
 import React, { useEffect, useMemo, useRef, useState } from "react";
 import { View, useWindowDimensions } from "react-native";
 import { FlashList } from "@shopify/flash-list";
-import TripCard from "./cards/TripCard";
-import PlaceCard from "./cards/PlaceCard";
-import UntrackedTimeCard from "./cards/UntrackedTimeCard";
-import { angularize, getAngularService } from "../angular-react-helper";
+import TripCard from "../cards/TripCard";
+import PlaceCard from "../cards/PlaceCard";
+import UntrackedTimeCard from "../cards/UntrackedTimeCard";
+import { angularize, getAngularService } from "../../angular-react-helper";
 import LoadMoreButton from "./LoadMoreButton";
-import useAppConfig from "../useAppConfig";
+import useAppConfig from "../../useAppConfig";
 import { useTranslation } from "react-i18next";
-import { invalidateMaps } from "../components/LeafletView";
+import { invalidateMaps } from "../../components/LeafletView";
 import { ActivityIndicator, Appbar } from "react-native-paper";
 import FilterSelect from "./FilterSelect";
 import DateSelect from "./DateSelect";
