+++ www/js/survey/multilabel/MultiLabelButtonGroup.tsx
@@ -9,7 +9,7 @@ import { View } from "react-native";
 import { IconButton, Text, useTheme } from "react-native-paper";
 import DiaryButton from "../../diary/DiaryButton";
 import { useTranslation } from "react-i18next";
-import { TimelineScrollContext } from "../../diary/TimelineScrollList";
+import { TimelineScrollContext } from "../../diary/list/TimelineScrollList";
 
 const MultilabelButtonGroup = ({ trip }) => {
   const { colors } = useTheme();
